Split a library out
