+++ psc-pages.cabal
@@ -10,12 +10,14 @@ category:            Language
 build-type:          Simple
 cabal-version:       >=1.10
 
-executable psc-pages
-  main-is:             Main.hs
-  other-modules:       
-  other-extensions:    
-  build-depends:       base >=4.7 && <4.8, 
-                       cheapskate -any, 
+library
+  hs-source-dirs:      src
+  ghc-options:         -Wall
+  default-language:    Haskell2010
+  buildable:           True
+  exposed:             True
+  build-depends:       base >=4.7 && <4.8,
+                       cheapskate -any,
                        purescript >= 0.6.9.3 && < 0.7,
                        blaze-markup,
                        blaze-html,
@@ -28,6 +30,26 @@ executable psc-pages
                        directory,
                        file-embed,
                        split
-  hs-source-dirs:      src
+  exposed-modules:     PscPages
+
+executable psc-pages
+  main-is:             Main.hs
+  other-modules:
+  other-extensions:
+  build-depends:       base >=4.7 && <4.8,
+                       purescript >= 0.6.9.3 && < 0.7,
+                       psc-pages -any,
+                       blaze-markup,
+                       blaze-html,
+                       transformers,
+                       mtl,
+                       optparse-applicative,
+                       text,
+                       data-default,
+                       filepath,
+                       directory,
+                       file-embed,
+                       split
+  hs-source-dirs:      psc-pages
   ghc-options:         -Wall
   default-language:    Haskell2010
